+++ source/test/matrix_test.cpp
@@ -324,7 +324,7 @@ BOOST_AUTO_TEST_CASE(matrix_vector_row_normalize)
 
     matrix_type normalized
     {
-        {1.0f / sqrtf(2.0f), 1.0f / sqrtf(2.0f)}
+        {1.0f / sqrt(2.0f), 1.0f / sqrt(2.0f)}
     };
     BOOST_CHECK(m.row(0).safe_normalize());
     KISMET_CHECK_APPROX_COLLECTIONS(m.row(0), normalized.row(0));
@@ -348,8 +348,8 @@ BOOST_AUTO_TEST_CASE(matrix_vector_column_normalize)
 
     matrix_type normalized
     {
-        { 1.0f / sqrtf(2.0f) }, 
-        { 1.0f / sqrtf(2.0f) }
+        { 1.0f / sqrt(2.0f) }, 
+        { 1.0f / sqrt(2.0f) }
     };
     BOOST_CHECK(m.column(0).safe_normalize());
     KISMET_CHECK_APPROX_COLLECTIONS(m.column(0), normalized.column(0));
